refactor #13: home/stay 이벤트 리스너 패키지 이동
