+++ backend/src/main/java/backend/capstone/domain/ongoinghomestatus/event/HomeStatusAnalysisEventListener.java
@@ -1,6 +1,7 @@
-package backend.capstone.domain.dayroute.event;
+package backend.capstone.domain.ongoinghomestatus.event;
 
-import backend.capstone.domain.ongoingstay.service.StayAnalysisService;
+import backend.capstone.domain.dayroute.event.GpsPointsUploadedEvent;
+import backend.capstone.domain.ongoinghomestatus.service.HomeStatusAnalysisService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Async;
@@ -11,17 +12,17 @@ import org.springframework.transaction.event.TransactionalEventListener;
 @Component
 @RequiredArgsConstructor
 @Slf4j
-public class StayAnalysisEventListener {
+public class HomeStatusAnalysisEventListener {
 
-    private final StayAnalysisService stayAnalysisService;
+    private final HomeStatusAnalysisService homeStatusAnalysisService;
 
-    @Async("stayAnalysisExecutor")
+    @Async("homeStatusAnalysisExecutor")
     @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
     public void handle(GpsPointsUploadedEvent event) {
         try {
-            stayAnalysisService.analyzeStay(event.dayRouteId());
+            homeStatusAnalysisService.analyzeHomeStatus(event.dayRouteId());
         } catch (Exception e) {
-            log.error("Stay analysis failed. dayRouteId={}", event.dayRouteId(), e);
+            log.error("Home status analysis failed. dayRouteId={}", event.dayRouteId(), e);
         }
     }
 }
